requirements.txt
numpy==1.13.3
olefile==0.44
pandas==0.21.0
Pillow==4.3.0
python-dateutil==2.6.1
pytz==2017.3
six==1.11.0
